+++ reproc/include/reproc/sink.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include <reproc/export.h>
+#include <reproc/reproc.h>
 
 #include <stdbool.h>
 #include <stdint.h>
@@ -12,17 +13,17 @@ extern "C" {
 #endif
 
 /*!
-Stores the output of a process in a single contiguous C string.
+Stores the output (both stdout and stderr) of a process in a single contiguous C
+string.
 
 Expects a `char **` with its value set to `NULL` as its initial context.
 (Re)Allocates memory as necessary to store the output and assigns it as the
 value of the given context. If allocating more memory fails, the already
 allocated memory is freed and the value of the given context is set to `NULL`.
 
-When `reproc_drain` finished after being called with `reproc_sink_string`, the
-value of the given context will either point to valid memory or will be set to
-`NULL`. This means it is always safe to call `free` on the context value after
-`reproc_drain` finishes.
+After calling `reproc_drain` with `reproc_sink_string`, the value of `context`
+will either point to valid memory or will be set to `NULL`. This means it is
+always safe to call `free` on `context`'s value after `reproc_drain` finishes.
 
 Because the context this function expects does not store the output size,
 `strlen` is called each time data is read to calculate the current size of the
@@ -35,14 +36,18 @@ their output because `strlen` is used to calculate the current output size.
 The `drain` example shows how to use `reproc_sink_string`.
 ```
 */
-REPROC_EXPORT bool
-reproc_sink_string(const uint8_t *buffer, unsigned int size, void *context);
+REPROC_EXPORT bool reproc_sink_string(REPROC_STREAM stream,
+                                      const uint8_t *buffer,
+                                      unsigned int size,
+                                      void *context);
 
 /*!
 Discards the output of a process.
 */
-REPROC_EXPORT bool
-reproc_sink_discard(const uint8_t *buffer, unsigned int size, void *context);
+REPROC_EXPORT bool reproc_sink_discard(REPROC_STREAM stream,
+                                       const uint8_t *buffer,
+                                       unsigned int size,
+                                       void *context);
 
 #ifdef __cplusplus
 }